requirements.txt
# Django and REST Framework
Django>=5.0,<6.0
djangorestframework>=3.14.0

# JWT Authentication
djangorestframework-simplejwt>=5.3.0
PyJWT>=2.8.0

# Image handling
Pillow>=10.0.0

# Environment variables
python-dotenv>=1.0.0

# Database (SQLite is built-in, but adding for explicit reference)
# For production, you might want:
# psycopg2-binary>=2.9.9  # PostgreSQL

# CORS headers (for frontend integration)
django-cors-headers>=4.3.0

# Development utilities
black>=23.0.0
flake8>=6.0.0
